requirements.txt
requests==2.18.4
Pillow==5.4.1
